Object Oriented Programming/Jar/test_jar.py
from jar import Jar


def main():
    test_str()


def test_init():
    jar = Jar()
    assert jar.capacity == 12
    jar_2 = Jar(3)
    assert jar_2.capacity == 3


def test_str():
    jar = Jar()
    assert str(jar) == ""
    jar.deposit(1)
    assert str(jar) == "🍪"
    jar.deposit(11)
    assert str(jar) == "🍪🍪🍪🍪🍪🍪🍪🍪🍪🍪🍪🍪"


def test_deposit():
    jar = Jar()
    jar.deposit(6)
    assert jar.size == 6
    jar.deposit(3)
    assert jar.size == 9
    jar.deposit(3)
    assert jar.size == 12


def test_withdraw():
    jar = Jar()
    jar.deposit(6)
    jar.withdraw(3)
    assert jar.size == 3
    jar.withdraw(3)
    assert jar.size == 0
